Export nimation is partly implemented
