+++ FireRender.Maya.Src/FireRender.vcxproj
@@ -149,6 +149,7 @@
     <ClCompile Include="..\RadeonProRenderSDK\RadeonProRender\inc\Math\half.cpp" />
     <ClCompile Include="..\RadeonProRenderSDK\RadeonProRender\rprTools\RPRStringIDMapper.cpp" />
     <ClCompile Include="..\RadeonProRenderSDK\RadeonProRender\rprTools\RprTools.cpp" />
+    <ClCompile Include="AnimationExporter.cpp" />
     <ClCompile Include="athenaCmd.cpp" />
     <ClCompile Include="athenaSystemInfo_Win.cpp" />
     <ClCompile Include="CompositeWrapper.cpp" />
@@ -301,6 +302,7 @@
     <ClInclude Include="..\RadeonProRenderSharedComponents\src\ImageFilter\ImageFilter.h" />
     <ClInclude Include="..\RadeonProRenderSharedComponents\src\Utils\Utils.h" />
     <ClInclude Include="..\RadeonProRenderSharedComponents\src\XMLMaterialExport\XMLMaterialExportCommon.h" />
+    <ClInclude Include="AnimationExporter.h" />
     <ClInclude Include="athenaCmd.h" />
     <ClInclude Include="athenaSystemInfo_Win.h" />
     <ClInclude Include="attributeNames.h" />
